Fixed minor warnings (signed/unsigned compare).
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@4402 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/tests/kits/app/bmessage/MessageBStringItemTest.h
@@ -130,7 +130,7 @@ struct TBStringAssertPolicy
 };
 bool TBStringAssertPolicy::Size(size_t size, BString& data)
 {
-	return size == data.Length() + 1;
+	return size == (size_t)data.Length() + 1;
 }
 //------------------------------------------------------------------------------
 typedef TMessageItemTest
